requirements.txt
blinker==1.8.2
bottle==0.13.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
proxy_tools==0.1.0
pycparser==2.22
pythonnet==3.0.4
pywebview==5.3.2
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
